+++ src/assets/config/config.ts
@@ -18,8 +18,7 @@ export const config: Config = {
     },
     enableRouterLoadingBar: true,
     auth: {
-      enabled: false,
-      backendAuthBaseURL: "https://api.sls.fi/"
+      enabled: false
     },
     openGraphMetaTags: {
       enabled: true,
@@ -33,7 +32,7 @@ export const config: Config = {
     prebuild: {
       sitemap: true,
       staticCollectionMenus: true,
-      featureBasedRoutes: false
+      featureBasedRoutes: true
     },
     ssr: {
       trustProxyHops: 2,
@@ -276,7 +275,7 @@ export const config: Config = {
         indexPlaces: true,
         indexWorks: true,
         search: true,
-        cookiePolicy: false,
+        cookiePolicy: true,
         termsOfUse: false,
         privacyPolicy: false,
         accessibilityStatement: false,